Enhancing code for better performance
