+++ messages/ar/homepage.json
@@ -1,6 +1,68 @@
 {
   "HomePage": {
-    "title": "مرحبا بالعالم!",
-    "about": "الذهاب إلى صفحة حول"
+    "entrytitle": {
+      "title1": "مرحبًا بك في ",
+      "title2": "منصة السوق",
+      "title3": "للعاملين المستقلين",
+      "title4": "المؤهلين!",
+      "subtitle1": "اعثر على أفضل المحترفين لأي مشروع أو قم بعرض",
+      "subtitle2": "مهاراتك من أجل تعاون ونجاح سلس.",
+      "slide2": "ابدأ بإنشاء ملفك الشخصي",
+      "slide3": "وكن المستقل المميز لهذا الشهر!"
+    },
+    "navigation": {
+      "nav1": "حول",
+      "nav2": "الخدمات",
+      "nav3": "الأسعار",
+      "nav4": "اتصل بنا"
+    },
+    "buttons": {
+      "login": "تسجيل الدخول",
+      "signup": "إنشاء حساب",
+      "getstarted": "ابدأ الآن"
+    },
+    "aboutus": {
+      "subtitle": "منصة آمنة وشاملة وسريعة",
+      "paragraphe": "مرحبًا بك في منصتنا، مساحة ديناميكية حيث يلتقي الإبداع بالفرصة. مهمتنا هي تمكين المستقلين من عرض مهاراتهم وخدماتهم في بيئة مهنية. في الوقت نفسه، نسعى لربط العملاء مع المواهب التي يمكنها تحقيق مشاريعهم. سواء كنت مستقلاً تسعى إلى تطوير مسارك المهني أو عميلاً يبحث عن الشريك المثالي، فإن منصتنا مصممة لتبسيط التواصل وضمان الشفافية وتعزيز التعاون. معًا، لنعيد تعريف كيفية إنجاز الأعمال!"
+    },
+    "services": {
+      "title": "خدمات مذهلة",
+      "subtitle": "اعثر بسهولة على المحترفين المؤهلين والخدمات المصممة حسب الطلب من خلال تطبيقنا. سواء كانت مشاريع شخصية أو احتياجات مهنية، تواصل مع أفضل العاملين المستقلين وأنجز عملك بكل بساطة.",
+      "title1": "فني صيانة",
+      "title2": "رسام مباني",
+      "title3": "نجار",
+      "title4": "بستاني/مصمم حدائق",
+      "title5": "ميكانيكي",
+      "description1": "ضمان حسن الأداء والسلامة على أساس يومي.",
+      "description2": "يقوم بتطبيق الطلاء على الجدران والأسطح، داخليًا وخارجيًا.",
+      "description3": "يصنع ويثبت الهياكل الخشبية للمباني والأسقف وما إلى ذلك.",
+      "description4": "يحافظ على المساحات الخضراء ويخلق الحدائق ويصمم المنتزهات.",
+      "description5": "فني يقوم بإصلاح وصيانة المركبات والآلات."
+    },
+    "pricing": {
+      "title1": "للعملاء",
+      "title2": "للعاملين المستقلين",
+      "subtitle1" : "يمكن للعملاء اكتشاف العاملين لحسابهم الخاص في جميع نقاط الأسعار، مع خيارات مخصصة",
+      "subtitle2" : "لجميع الميزانيات واحتياجات المشروع.",
+      "subtitle3" : "لجميع الميزانيات واحتياجات المشاريع، يتمتع المستقلون بحرية تحديد الأسعار الخاصة بهم والعثور على المشاريع",
+      "subtitle4" : "التي تتوافق مع مستوى مهارتهم وأهدافهم من 50 درهم."
+    },
+    "footer": {
+      "placeholder" : "أدخل بريدك الإلكتروني",
+      "subscribe": "اشترك",
+      "subscribenow": "اشترك الآن",
+      "Information": {
+        "information": "معلومات",
+        "testimonials": "الشهادات",
+        "events": "الفعاليات"
+      },
+      "HelpfulLinks": {
+        "title": "روابط مفيدة",
+        "services": "الخدمات",
+        "supports": "الدعم",
+        "terms&privacy": "الشروط والخصوصية"
+      },
+      "company": "2024 © company.Ltd | جميع الحقوق محفوظة"
+    }
   }
 }
