+++ messages/en/homepage.json
@@ -1,6 +1,73 @@
 {
   "HomePage": {
-    "title": "Hello world!",
-    "about": "Go to the about page"
-  }
+      "entrytitle": {
+          "title1": "Welcome To ",
+          "title2": "Your MarketPlace",
+          "title3": "For Skilled",
+          "title4": "Freelancers",
+          "subtitle1": "Find top professionals for any project, or showcase your",
+          "subtitle2": "skills for seamless collaboration and success.",
+          "slide2": "Start By Creating Your Profile",
+          "slide3": "And Be The Freelancer Of the Month !"
+      },
+      "navigation": {
+          "nav1" :"About us",
+          "nav2" :"Services",
+          "nav3" :"Pricing",
+          "nav4" :"Contact us"
+      },
+      "buttons": {
+          "login" : "Log In",
+          "signup": "Sign Up",
+          "getstarted": "Get Started"
+      },
+      "aboutus": {
+          "subtitle" :"Safe, comprehensive and fast platform",
+          "paragraphe" : "Welcome to our platform, a dynamic space where creativity meets opportunity. Our mission is to empower freelancers by providing them with a professional environment to showcase their skills and services. At the same time, we strive to connect clients with talented individuals who can bring their projects to life. Whether you're a freelancer looking to grow your career or a client searching for the perfect collaborator, our platform is designed to simplify connections, ensure transparency, and foster collaboration. Together, let's redefine the way work gets done!"
+      },
+      "services": {
+          "title" : "Amazing Services",
+          "subtitle" :"Find skilled professionals and tailored services easily with our app. Whether for personal projects or business needs, connect with top freelancers and get the job done seamlessly",
+          "title1" : "Maintenance agent",
+          "title2" : "House painter",
+          "title3" : "Carpenter",
+          "title4" : "Gardener/Landscaper",
+          "title5" : "Mechanic",
+          "description1": "Guaranteeing proper functioning and safety on a daily basis.",
+          "description2": "Applies paint to walls and surfaces, both indoors and outdoors.",
+          "description3": "Manufactures and installs wooden structures for buildings, roofs, etc.",
+          "description4": "Maintains green spaces, creates and develops gardens and parks.",
+          "description5": "Technician who repairs and maintains vehicles and machinery."
+      },
+      "pricing": {
+          "title1" : "For Clients",
+          "title2" : "For Freelancers",
+          "subtitle1" : "Clients can discover freelancers at every price point, with options to fit",
+          "subtitle2" : "any budget and project needs.",
+          "subtitle3" : "Freelancers have the freedom to set their own prices and find projects",
+          "subtitle4" : "that match their skill level and goals starting from 50DH."
+      },
+      "footer":{
+          "placeholder" : "Enter Your Email",
+          "subscribe" : "subscribe",
+          "subscribenow" : "Subscribe Now",
+          "Information": {
+            "information" : "Information",
+            "testimonials" : "Testimonials",
+            "events" : "Events"
+          },
+          "HelpfulLinks": {
+            "title" : "helpful links",
+            "services" : "Services",
+            "supports" : "Supports",
+            "terms&privacy" : "Terms & Privacy"
+          },
+          "company" : "2024 © company.Ltd | All Right reserved"
+          
+      }
+
+
+
+      
+}
 }
