+++ src/app/globals.css
@@ -8,7 +8,7 @@
   width: 0;
 }
 nav {
-  font-family: "Roboto", sans-serif;
+  font-family: "Inter", sans-serif;
   font-weight: 700;
   font-style: normal;
 }
@@ -22,9 +22,6 @@ nav {
   font-weight: 700;
 }
 
-.bri {
-  color: #009FFD;
-}
 img{
   object-fit: cover;
 }
