+++ src/components/ui/LanguageSwitcher.tsx
@@ -11,14 +11,15 @@ import {
 import {useTranslations} from "next-intl";
 import {Link, locales} from "@/i18n/routing";
 import {Languages} from "lucide-react";
+import { cn } from "@/lib/utils";
 
 
-export function LanguageSwitcher({locale = locales[0]}:{locale?: string}) {
+export function LanguageSwitcher({classname="",locale = locales[0]}:{locale?: string, classname?: string}) {
     const t = useTranslations('lang');
     return (
         <DropdownMenu>
             <DropdownMenuTrigger asChild>
-                <Button className={"w-6 h-6"} variant={"ghost"}>
+                <Button className={cn("w-6 h-6", classname)} variant={"ghost"}>
                     <Languages />
                     <p className={"sr-only"}>{t("title")}</p>
                 </Button>
