+++ src/components/ui/logo.tsx
@@ -33,52 +33,28 @@ const Logo = (
                     ? <Image
                         src={logo_path.dark[variant]}
                         alt={"Brikool"}
-                        width={width ?? 80}
-                        height={height ?? 80}/>
+                        width={width ?? 60}
+                        height={height ?? 60}/>
                     : <Image
                         src={logo_path.light[variant]}
                         alt={"Brikool"}
-                        width={width ?? 80}
-                        height={height ?? 80}/>
+                        width={width ?? 60}
+                        height={height ?? 60}/>
             }
         </>
     );
 }
 
 export {Logo}
-import React from 'react'
 
-export default function logo() {
-  return (
-  //   <svg
-  //   version="1.1"
-  //   id="Layer_1"
-  //   x="0px"
-  //   y="0px"
-  //   width="70.04"
-  //   height="70.91"
-  //   viewBox="0 0 149.36669 170.44168"
-  //   enableBackground="new 0 0 577 199"
-  //   xmlSpace="preserve"
-  // >
-  //   <path
-  //     fill="#343434"
-  //     opacity="1"
-  //     stroke="none"
-  //     d="m 30.959127,27.112813 c 3.9301,-2.284016 7.628926,-4.310486 11.5178,-6.558102 0.590046,-0.394291 0.990051,-0.567432 1.966053,-1.083751 9.738525,-5.688744 18.901054,-11.0343158 28.063583,-16.3798901 2.150815,-0.9079944 3.586456,0.497532 5.0466,1.330309 17.906333,10.2127391 35.788117,20.4707961 53.541367,30.9457251 1.58327,0.93417 3.04184,3.413838 3.15551,5.261998 0.38784,6.306503 0.29014,12.657112 0.0439,18.97789 -0.0581,1.491172 -1.03716,3.60337 -2.24321,4.294609 -18.74775,10.745109 -37.601854,21.30468 -56.442246,31.887796 -0.238632,0.13406 -0.6185,0.0167 -1.405685,0.0167 0,-7.63552 -0.130364,-15.230617 0.13784,-22.811618 0.04114,-1.162796 1.726563,-2.579522 2.978943,-3.338524 C 88.402956,62.938977 99.560277,56.343991 111.17641,49.426747 99.141567,42.592011 87.530726,35.930348 75.811495,29.465119 c -0.986954,-0.544483 -2.927681,-0.175426 -4.042572,0.450707 -14.638489,8.221103 -29.26265,16.472629 -43.711521,25.018947 -1.463711,0.865769 -2.486386,3.689263 -2.499154,5.61232 -0.164833,24.820603 -0.109066,49.642667 -0.109073,74.464367 0,1.77916 4e-6,3.55832 4e-6,6.07044 -6.961693,-3.97905 -13.323468,-7.49565 -19.5203468,-11.28212 -0.963316,-0.58862 -1.656262,-2.40836 -1.660639,-3.66164 -0.09368,-26.820437 -0.116911,-53.641546 0.04691,-80.461305 0.0094,-1.546082 1.35157,-3.691929 2.707857,-4.521519 C 14.818581,36.387013 22.805687,31.931767 30.959129,27.112813 Z"
-  //   />
-  //   <path
-  //     fill="#13ba83"
-  //     opacity="1"
-  //     stroke="none"
-  //     d="m 61.171305,78.924817 c -5.4e-5,17.659626 0.103538,34.820383 -0.08508,51.977943 -0.03907,3.55457 1.21463,5.38268 3.972878,7.64139 6.414062,5.25239 11.388523,4.22957 18.368189,0.11406 11.935876,-7.0379 23.708148,-13.67408 35.501048,-21.33167 2.86274,-1.85889 3.86783,-3.4904 4.06288,-7.02364 0.12615,-2.28518 0.12581,-4.30021 -0.17084,-6.46053 -0.25339,-1.84526 -1.5567,-3.882137 -3.02362,-4.911114 -3.37498,-2.367423 -7.14601,-4.170239 -11.35403,-6.538599 7.04735,-4.2184 13.5526,-8.181524 20.15568,-11.97431 0.79604,-0.457237 2.33315,-0.365074 3.16585,0.107895 4.04816,2.299301 8.04341,4.712303 11.88854,7.332785 0.86384,0.58872 1.47416,2.16649 1.48446,3.29547 0.10639,11.660903 0.15378,23.324683 -0.0473,34.983093 -0.0248,1.43734 -1.24321,3.45525 -2.4995,4.18076 -21.7527,12.56227 -43.569663,25.01474 -65.484948,37.2906 -1.512116,0.84702 -4.293961,0.8304 -5.831291,0.002 -10.102105,-5.44445 -20.07795,-11.13241 -29.943653,-16.99557 -1.36895,-0.81357 -2.69825,-3.02406 -2.708066,-4.59624 -0.165443,-26.48811 -0.148105,-52.977643 -0.0352,-79.466284 0.0059,-1.391754 0.84496,-3.411911 1.937695,-4.063324 6.534298,-3.895286 13.256916,-7.47467 20.646336,-11.563965 v 27.99939 z"
-  //   />
-  // </svg>
+// import React from 'react'
+// export default function logo() {
+//   return (
 
-  <svg width="60" height="61" viewBox="0 0 90 103" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M18.5756 16.2675C20.9337 14.8971 23.153 13.6812 25.4863 12.3326C25.8403 12.096 26.0803 11.9921 26.6659 11.6824C32.5091 8.26911 38.0066 5.06177 43.5041 1.85442C44.7946 1.30963 45.656 2.15294 46.5321 2.65261C57.2759 8.78025 68.0049 14.9351 78.6569 21.22C79.6068 21.7805 80.482 23.2683 80.5502 24.3772C80.7829 28.1611 80.7243 31.9715 80.5765 35.764C80.5417 36.6587 79.9542 37.926 79.2306 38.3407C67.9819 44.7878 56.6695 51.1235 45.3652 57.4734C45.2221 57.5538 44.9941 57.4834 44.5218 57.4834C44.5218 52.9021 44.4436 48.3451 44.6045 43.7965C44.6292 43.0988 45.6405 42.2487 46.3919 41.7933C53.0419 37.7632 59.7363 33.8062 66.706 29.6558C59.4851 25.555 52.5186 21.558 45.4871 17.6789C44.8949 17.3522 43.7304 17.5736 43.0615 17.9493C34.2784 22.8819 25.5039 27.8329 16.8346 32.9606C15.9564 33.4801 15.3428 35.1742 15.3351 36.328C15.2362 51.2204 15.2697 66.1136 15.2697 81.0066C15.2697 82.0741 15.2697 83.1416 15.2697 84.6489C11.0927 82.2615 7.27559 80.1515 3.55746 77.8796C2.97947 77.5265 2.5637 76.4346 2.56108 75.6827C2.50487 59.5904 2.49093 43.4977 2.58922 27.4059C2.59486 26.4782 3.40017 25.1907 4.21394 24.693C8.89131 21.832 13.6836 19.1588 18.5756 16.2675Z" fill="#333333"/>
-<path d="M36.7026 47.3547C36.7026 57.9505 36.7648 68.2469 36.6516 78.5415C36.6281 80.6742 37.3804 81.7711 39.0353 83.1263C42.8837 86.2777 45.8684 85.6641 50.0562 83.1947C57.2177 78.972 64.2811 74.9903 71.3568 70.3958C73.0745 69.2804 73.6775 68.3015 73.7946 66.1816C73.8703 64.8105 73.8701 63.6014 73.6921 62.3053C73.54 61.1981 72.758 59.976 71.8779 59.3586C69.8529 57.9381 67.5903 56.8564 65.0655 55.4354C69.2939 52.9044 73.197 50.5265 77.1589 48.2508C77.6365 47.9765 78.5588 48.0318 79.0584 48.3156C81.4873 49.6952 83.8844 51.143 86.1915 52.7152C86.7098 53.0685 87.076 54.0151 87.0822 54.6925C87.146 61.6891 87.1745 68.6873 87.0538 75.6824C87.0389 76.5448 86.3079 77.7555 85.5541 78.1908C72.5025 85.7282 59.4123 93.1997 46.2632 100.565C45.3559 101.073 43.6868 101.063 42.7644 100.566C36.7031 97.2997 30.7176 93.887 24.7982 90.3691C23.9768 89.8809 23.1792 88.5546 23.1734 87.6113C23.0741 71.7184 23.0845 55.8247 23.1522 39.9315C23.1558 39.0965 23.6592 37.8844 24.3148 37.4936C28.2354 35.1564 32.269 33.0088 36.7026 30.5552L36.7026 47.3547Z" fill="#009FFD"/>
-</svg>
+//   <svg width="60" height="61" viewBox="0 0 90 103" fill="none" xmlns="http://www.w3.org/2000/svg">
+// <path d="M18.5756 16.2675C20.9337 14.8971 23.153 13.6812 25.4863 12.3326C25.8403 12.096 26.0803 11.9921 26.6659 11.6824C32.5091 8.26911 38.0066 5.06177 43.5041 1.85442C44.7946 1.30963 45.656 2.15294 46.5321 2.65261C57.2759 8.78025 68.0049 14.9351 78.6569 21.22C79.6068 21.7805 80.482 23.2683 80.5502 24.3772C80.7829 28.1611 80.7243 31.9715 80.5765 35.764C80.5417 36.6587 79.9542 37.926 79.2306 38.3407C67.9819 44.7878 56.6695 51.1235 45.3652 57.4734C45.2221 57.5538 44.9941 57.4834 44.5218 57.4834C44.5218 52.9021 44.4436 48.3451 44.6045 43.7965C44.6292 43.0988 45.6405 42.2487 46.3919 41.7933C53.0419 37.7632 59.7363 33.8062 66.706 29.6558C59.4851 25.555 52.5186 21.558 45.4871 17.6789C44.8949 17.3522 43.7304 17.5736 43.0615 17.9493C34.2784 22.8819 25.5039 27.8329 16.8346 32.9606C15.9564 33.4801 15.3428 35.1742 15.3351 36.328C15.2362 51.2204 15.2697 66.1136 15.2697 81.0066C15.2697 82.0741 15.2697 83.1416 15.2697 84.6489C11.0927 82.2615 7.27559 80.1515 3.55746 77.8796C2.97947 77.5265 2.5637 76.4346 2.56108 75.6827C2.50487 59.5904 2.49093 43.4977 2.58922 27.4059C2.59486 26.4782 3.40017 25.1907 4.21394 24.693C8.89131 21.832 13.6836 19.1588 18.5756 16.2675Z" fill="#333333"/>
+// <path d="M36.7026 47.3547C36.7026 57.9505 36.7648 68.2469 36.6516 78.5415C36.6281 80.6742 37.3804 81.7711 39.0353 83.1263C42.8837 86.2777 45.8684 85.6641 50.0562 83.1947C57.2177 78.972 64.2811 74.9903 71.3568 70.3958C73.0745 69.2804 73.6775 68.3015 73.7946 66.1816C73.8703 64.8105 73.8701 63.6014 73.6921 62.3053C73.54 61.1981 72.758 59.976 71.8779 59.3586C69.8529 57.9381 67.5903 56.8564 65.0655 55.4354C69.2939 52.9044 73.197 50.5265 77.1589 48.2508C77.6365 47.9765 78.5588 48.0318 79.0584 48.3156C81.4873 49.6952 83.8844 51.143 86.1915 52.7152C86.7098 53.0685 87.076 54.0151 87.0822 54.6925C87.146 61.6891 87.1745 68.6873 87.0538 75.6824C87.0389 76.5448 86.3079 77.7555 85.5541 78.1908C72.5025 85.7282 59.4123 93.1997 46.2632 100.565C45.3559 101.073 43.6868 101.063 42.7644 100.566C36.7031 97.2997 30.7176 93.887 24.7982 90.3691C23.9768 89.8809 23.1792 88.5546 23.1734 87.6113C23.0741 71.7184 23.0845 55.8247 23.1522 39.9315C23.1558 39.0965 23.6592 37.8844 24.3148 37.4936C28.2354 35.1564 32.269 33.0088 36.7026 30.5552L36.7026 47.3547Z" fill="#009FFD"/>
+// </svg>
 
-  )
-}
+//   )
+// }
