+++ src/config/navigation/static-navlinks.ts
@@ -6,5 +6,17 @@ export const staticNavLinks = {
     termsAndPrivacy: {
         href: '/terms-and-privacy',
         label: 'Terms and Privacy',
+    },
+    services: {
+        title: "Services",
+        href: "#services"
+    },
+    pricing: {
+        title: "Pricing",
+        href: "#pricing"
+    },
+    contactus: {
+        title: "Contact us",
+        href: "#footer"
     }
 }
